Fix & tối ưu các api get ds quản lý, Thêm vài api thống kê

+++ backend/src/main/java/com/megacart/dto/response/DonHangGiaoHangResponse.java
@@ -11,6 +11,6 @@ public class DonHangGiaoHangResponse {
     private String tenNguoiNhan;
     private String sdtNhanHang;
     private String diaChiNhanHang;
-    private int tongTien;
+    private long tongTien;
     private TrangThaiThanhToan trangThaiThanhToan;
 }
